+++ codex-rs/login/src/lib.rs
@@ -25,6 +25,11 @@ use crate::token_data::parse_id_token;
 
 mod token_data;
 mod server;
+mod pkce;
+mod jwt_utils;
+mod auth_file;
+mod success_url;
+mod redeem;
 pub use server::LoginServerOptions;
 pub use server::run_local_login_server_with_options;
 pub use server::process_callback_headless;
